backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-search-results>=2.4.2
google-generativeai>=0.4.0
icalendar>=5.0.11
pydantic>=2.8,<2.9
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
fpdf2>=2.7.0
